+++ components/CardMessages.js
@@ -1,0 +1,10 @@
+
+export default function Card({ children }) {
+    return (
+       <div className="bg-gray-100 shadow-md rounded-md p-4 w-full">
+          {children}
+       </div>
+    );
+ }
+ 
+
