+++ pages/directmessage.js
@@ -1,174 +1,23 @@
-import { Component } from 'react';
-
-class SendButton extends Component {
-  render() {
-    return (<div className="send_message" onClick={this.props.handleClick}>
-      <div className="text text-center">
-        <button className='btn '>Send</button>
-      </div>
-    </div>);
-  }
-}
-
-class MessageTextBoxContainer extends Component {
-  render() {
-    return (
-      <div className="rounded text-center">
-        <input id="msg_input" className="message_input" placeholder="Type your messages here..." value={this.props.message} onChange={this.props.onChange} onKeyPress={this.props._handleKeyPress} />
-      </div>
-    );
-  }
-}
-
-class Avartar extends Component {
-  render() {
-    return (
-      <div className="avatar" />
-    );
-  }
-}
-
-class BotMessageBox extends Component {
-  constructor(props) {
-    super(props);
-  }
-  render() {
-    return (
-      <li className="message left appeared">
-        <Avartar></Avartar>
-        <div className="text_wrapper">
-          <div className="text text-center color: black">{this.props.message}</div>
+import CardMessages from "../components/CardMessages"
+export default function directmessage() {
+  return (
+    <div className="container relative m-3 h-full">
+      <div className="absolute bottom-x-0 bottom-0 w-full">
+        <div className="flex w-11/12 justify-center items-center">
+          <CardMessages>
+          <div className="w-full inline-flex">
+              <textarea className="textarea textarea-bordered w-11/12" placeholder="Send a Message..."></textarea>
+                <button className="btn btn my-3 mx-2">
+                Send
+              </button>
+              <button className="btn btn my-3">
+                Images
+              </button>
+          </div>
+          </CardMessages>    
         </div>
-      </li>
-    );
-  }
-}
-
-class UserMessageBox extends Component {
-  constructor(props) {
-    super(props);
-
-  }
-  render() {
-    return (
-      <li className={`message ${this.props.appearance} appeared`}>
-        <Avartar></Avartar>
-        <div className="text_wrapper">
-          <div className="text">{this.props.message}</div>
-        </div>
-      </li>
-    );
-  }
-}
-
-class MessagesContainer extends Component {
-  constructor(props) {
-    super(props);
-    this.createBotMessages = this.createBotMessages.bind(this);
-  }
-
-  scrollToBottom = () => {
-    var el = this.refs.scroll;
-    el.scrollTop = el.scrollHeight;
-  }
-
-  componentDidMount() {
-    this.scrollToBottom();
-  }
-
-  componentDidUpdate() {
-    this.scrollToBottom();
-  }
-
-  createBotMessages() {
-    console.log(this.props.messages);
-    return this.props.messages.map((message, index) =>
-      <UserMessageBox key={index} message={message["message"]} appearance={message["isbotmessage"] ? "left" : "right"} />
-    );
-  }
-
-  render() {
-
-    return (
-      <ul className="messages text-center" ref="scroll">
-        {this.createBotMessages()}
-      </ul>
-    );
-  }
-}
-
-
-class directmessage extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { "messages": [], "current_message": "" }
-    this.handleClick = this.handleClick.bind(this);
-    this._handleKeyPress = this._handleKeyPress.bind(this);
-    this.onChange = this.onChange.bind(this);
-    this.addMessageBox = this.addMessageBox.bind(this);
-  }
-
-
-  addMessageBox(enter = true) {
-    let messages = this.state.messages;
-    let current_message = this.state.current_message;
-    console.log(this.state);
-    if (current_message && enter) {
-      messages = [...messages, { "message": current_message }];
-      fetch("http://localhost:5000?message=" + current_message)
-        .then(res => res.json())
-        .then(
-          (result) => {
-            console.log(result);
-            this.setState({
-              messages: [...messages, { "message": result["message"], "isbotmessage": true }]
-            });
-          },
-          (error) => {
-            //do nothing for now
-          }
-        );
-      current_message = ""
-    }
-    this.setState({
-      current_message: current_message,
-      messages
-    });
-
-  }
-
-  handleClick() {
-    this.addMessageBox();
-  }
-
-  onChange(e) {
-    this.setState({
-      current_message: e.target.value
-    });
-  }
-
-  _handleKeyPress(e) {
-    let enter_pressed = false;
-    if (e.key === "Enter") {
-      enter_pressed = true;
-    }
-    this.addMessageBox(enter_pressed)
-  }
-
-  render() {
-    return (
-      <div className="chat_window">
-        <MessagesContainer messages={this.state.messages}></MessagesContainer>
-        <div className="bottom_wrapper clearfix">
-          <MessageTextBoxContainer
-            _handleKeyPress={this._handleKeyPress}
-            onChange={this.onChange}
-            message={this.state.current_message}></MessageTextBoxContainer>
-          <SendButton handleClick={this.handleClick}></SendButton>
-        </div>
-
+        
       </div>
-    );
-  }
+    </div>
+  )
 }
-export default directmessage;
